requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
jinja2==3.1.2

# Security scanning integrations
python-owasp-zap-v2.4==0.0.21
xmltodict==0.13.0
defusedxml==0.7.1

# LLM integrations
openai==1.12.0
anthropic==0.18.1
transformers==4.37.2
torch==2.2.0
accelerate==0.26.1
bitsandbytes==0.42.0
sentence-transformers==2.3.1

# Local LLM support
ollama==0.1.6
llama-cpp-python==0.2.55

# NLP and evaluation
nltk==3.8.1
rouge-score==0.1.2
sacrebleu==2.4.0
textstat==0.7.3
spacy==3.7.2

# Data processing
pandas==2.2.0
numpy==1.26.3
jsonschema==4.21.1

# Web scraping and APIs
beautifulsoup4==4.12.3
lxml==5.1.0

# CLI and logging
click==8.1.7
rich==13.7.0
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Dashboard
Flask==3.0.0
pytest-asyncio==0.21.1

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# CI/CD integrations
python-gitlab==4.4.0
PyGithub==2.1.1
jenkins-job-builder==6.0.0

# Cloud providers
boto3==1.34.34  # AWS
azure-devops==7.1.0b4  # Azure

# Reporting and visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
fpdf2==2.7.7

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
